Estimate the coefficients for the recommended FAO equation

Estimate the coefficients for the recommended FAO equation using: ambient temperature and dew point taken two hours after sunset, and minimum daily temperature, which were taken in nights with radiative frost (wind less than 2 m/s^2, no clouds, no fog, no rain).

buildFAO(dw, temp, tmin)

Arguments

dw
[°C]: an array of dew points, two hours after sunset.

temp
[°C]: an array of ambient temperature, two hours after sunset.

tmin
[°C]: minimum temperature

Value

A FAOFrostModel object with a, b, and c values, which can be used to estimate minimum temperature (Tmin) using temp and dw. This function also returns Tp (predicted temperature using the equation), Rp (residuals, the difference between tmin given and Tp) and r2 which is the coefficient of correlation (R squared).

Details

The method was extracted from the documentation and previous implementation in FFST.xls file, which is name in the book "Frost Protection: fundamentals, practice, and economics. Volume 1. Authors: Richard L Snyder, J. Paulo de Melo-Abreu. Food and Agriculture Organization of the United Nations. 2005"

This function implements the method, resolve the equation and find the coefficients. Equation: Tmim = a * temp + b * dw + c where "temp" and "dw" are the temperature and dew point respectively, which must be taken two hours after sunset

For more details please check: <http://www.fao.org/docrep/008/y7223e/y7223e0b.htm#bm11.8> <http://www.fao.org/docrep/008/y7223e/y7223e0b.htm> FFST spreasheet <http://biomet.ucdavis.edu/frostprotection/FTrend/FFST_FTrend.htm>

Examples

x1 <- rnorm(100,mean=2,sd=5)
x2 <- rnorm(100,mean=1,sd=3)
y <- rnorm(100,mean=0,sd=2)
buildFAO(dw = x2,temp=x1,tmin=y)

#> An object of class "FAOFrostModel"
#> Slot "a":
#> [1] 0.04049204
#> 
#> Slot "b":
#> [1] 0.1652381
#> 
#> Slot "c":
#> [1] -0.3775852
#> 
#> Slot "Tp":
#>   [1] -0.60676339 -0.46902162 -1.14865147 -0.52685219 -0.87613412 -0.24116481
#>   [7] -0.77708582 -0.44746819 -0.06821350  0.29643453 -0.34701118 -0.69772691
#>  [13]  0.41488306 -0.68057974  0.25424620  0.54576152 -0.02644132 -0.97693521
#>  [19]  0.10220423  0.21539452  0.97436420  0.16492090  0.29446100  0.22946253
#>  [25]  0.30431508 -0.14003990 -0.02161621 -0.26479191 -0.56062290 -0.34231251
#>  [31] -0.51017670  0.32383480 -0.04151676 -0.44962407 -0.57708613  0.06673160
#>  [37] -0.94201580 -0.33011546  0.24882198  0.11808820 -0.19782468 -0.67086950
#>  [43]  0.78873181  0.19765702  0.44446931  1.08172523 -0.67699889  0.24069463
#>  [49] -0.85551558  0.24139157  0.25590975 -0.12008032 -0.73721645 -0.40409635
#>  [55]  0.39638032  0.31589272 -0.57269598  0.23508490  0.18357112  0.31059144
#>  [61] -0.85573799  0.19786478 -0.05903316 -0.05074129 -1.25739469 -0.71968508
#>  [67] -0.47738627  0.07556288 -0.17415380 -0.21881810 -0.48014186 -0.55669405
#>  [73]  0.76928166  0.39287304  0.51660519  0.48823363 -0.24343322 -0.78582230
#>  [79] -0.24800355  0.94675513 -0.25812991 -0.43683835 -0.81024623  0.73767996
#>  [85]  0.03235444  1.08148945 -0.73665660  1.10832914  0.48293063  0.35775733
#>  [91]  0.51162114 -0.11292642 -0.32714271 -0.42603867  0.84139848  0.27325533
#>  [97] -0.01647090  0.62470851  0.75055385  0.11200147
#> 
#> Slot "Rp":
#>   [1] -0.25199679  3.18994427  1.00693661 -0.01745517 -4.01722594  0.37213809
#>   [7] -1.41993198 -0.81888817 -4.05909540  5.00142952 -1.95978559  0.01645115
#>  [13]  1.15784209 -1.86044648  0.83003690 -0.39554972  1.14347017  1.80774801
#>  [19] -3.00680377  1.66701773 -1.65223595 -0.31606939 -0.21405222  0.01913960
#>  [25] -2.30118019  2.60682003  0.70246518 -0.68061305  1.97812902 -2.71560492
#>  [31]  0.98502739 -2.94946330  1.53557393 -2.67541280  0.71919285 -1.34580114
#>  [37] -0.74837567  1.68060486  2.05792961 -3.49109768 -1.60780522  3.30613689
#>  [43]  1.41164767  2.20987866 -3.30701087  1.68409649  0.68325077 -0.39646828
#>  [49]  1.73837203  0.01645423 -1.91633827 -0.88710550 -1.65006592 -1.09935030
#>  [55]  2.51530248 -1.97309979  1.15224490 -1.19519187 -1.39322983  2.60962893
#>  [61]  1.15509670 -3.06450698  0.03842653 -0.37373078 -0.55528567 -3.48461988
#>  [67]  4.26410720 -2.01181456 -0.03105228  0.69873725  0.60193964 -3.79845800
#>  [73] -1.00500195 -0.16828347 -0.50083279  3.26725411  4.56094633  2.20525135
#>  [79]  1.78197031 -1.56317797  2.28213360 -1.40126484  1.93700638 -0.09271446
#>  [85]  0.70099429  1.17818086 -1.14633955 -0.67265386  2.34789396 -1.12522343
#>  [91] -0.85979389 -0.33056262 -1.69191474  1.38748921  2.36741617 -3.30330439
#>  [97] -2.81557693  1.12884614  0.49771097  4.11255310
#> 
#> Slot "r2":
#> [1] 0.06615118
#> 

#data example taken from FAO Book
t0 <- c(3.2,0.8,0.2,2.6,4.4,5.2,2.7,1.2,4.5,5.6) # temperature 2 hours after sunset
td <- c(-4.2,-8.8,-6.5,-6.2,-6.1,2.6,-0.7,-1.7,-1.2,0.1) # dew point 2 hours after sunset
tn <- c(-3.1,-5,-6.3,-5.4,-4,-2.5,-4.8,-5,-4.4,-3.3)
buildFAO(dw = td,temp=t0,tmin=tn)

#> An object of class "FAOFrostModel"
#> Slot "a":
#> [1] 0.5028121
#> 
#> Slot "b":
#> [1] 0.02647164
#> 
#> Slot "c":
#> [1] -5.821987
#> 
#> Slot "Tp":
#>  [1] -4.324169 -5.652687 -5.893490 -4.678799 -3.771090 -3.138537 -4.482924
#>  [8] -5.263614 -3.591098 -3.003591
#> 
#> Slot "Rp":
#>  [1]  1.2241687  0.6526874 -0.4065101 -0.7212007 -0.2289098  0.6385372
#>  [7] -0.3170760  0.2636139 -0.8089020 -0.2964085
#> 
#> Slot "r2":
#> [1] 0.7587214
#>