Estimate the coefficients for the recommended FAO equation

Estimate the coefficients for the recommended FAO equation using temperature two hours after sunset and minimum temperature.

buildFAOTemp(temp, tmin)

Arguments

temp
[°C]: an array of ambient temperature, two hours after sunset.

tmin
[°C]: minimum temperature

Value

A FAOFrostModel object with a, b, and c values, which can be used to estimate minimum temperature (Tmin) using temp and dw. This function also returns Tp (predicted temperature using the equation), Rp (residuals, the difference between tmin given and Tp) and r2 which is the coefficient of correlation (R squared).

Details

The method was extracted from the documentation and previous implementation in FFST.xls file, which is name in the book "Frost Protection: fundamentals, practice, and economics. Volume 1. Authors: Richard L Snyder, J. Paulo de Melo-Abreu. Food and Agriculture Organization of the United Nations. 2005"

This function implements the method, resolve the equation and find the coefficients. Equation: Tmim = a * temp + c where "temp" is the temperature which must be taken two hours after sunset

For more details please check: <http://www.fao.org/docrep/008/y7223e/y7223e0b.htm#bm11.8> <http://www.fao.org/docrep/008/y7223e/y7223e0b.htm> FFST spreasheet <http://biomet.ucdavis.edu/frostprotection/FTrend/FFST_FTrend.htm>

Examples

x1 <- rnorm(100,mean=2,sd=5)
x2 <- rnorm(100,mean=1,sd=3)
y <- rnorm(100,mean=0,sd=2)
buildFAO(dw = x2,temp=x1,tmin=y)

#> An object of class "FAOFrostModel"
#> Slot "a":
#> [1] 0.02695276
#> 
#> Slot "b":
#> [1] -0.05187522
#> 
#> Slot "c":
#> [1] -0.1583918
#> 
#> Slot "Tp":
#>   [1] -0.474537557 -0.168270473  0.138440008 -0.155475584  0.103198918
#>   [6] -0.025933430 -0.171406827 -0.217325418  0.193135940 -0.116058368
#>  [11] -0.227498666 -0.211793346  0.023182334 -0.327062881 -0.308694198
#>  [16] -0.249748820 -0.077428061  0.081573730 -0.437784025 -0.164284003
#>  [21] -0.159083073 -0.082340844 -0.072557052  0.008572697  0.587586518
#>  [26] -0.467196644  0.007234603  0.326270803 -0.119683133 -0.204372260
#>  [31] -0.085196212 -0.290740461 -0.081630351 -0.358488035  0.001081551
#>  [36]  0.090436517 -0.127896992 -0.234137785 -0.159489497  0.233732957
#>  [41]  0.031122976  0.014482219  0.021797587 -0.286745645 -0.382212424
#>  [46]  0.123599159  0.033171128 -0.054990479 -0.408058485 -0.104361861
#>  [51] -0.511309073 -0.236129446  0.057558531  0.323825460  0.051946773
#>  [56]  0.372726481 -0.208667024  0.145235892 -0.409085574 -0.238611113
#>  [61] -0.361152281  0.027917262 -0.304902394  0.207150521 -0.026732508
#>  [66]  0.018917715 -0.294904423 -0.249456763 -0.191659711 -0.279547992
#>  [71] -0.034283978 -0.124780203 -0.151880947  0.049775986  0.012126113
#>  [76]  0.075677235 -0.011248539 -0.193799612 -0.249951887  0.216461461
#>  [81] -0.070724118 -0.304664477 -0.286183583 -0.291204048 -0.336098880
#>  [86] -0.050934094 -0.238954100 -0.147139123 -0.543756546  0.209707860
#>  [91] -0.221947401  0.010806823  0.161264855 -0.031804407 -0.076402834
#>  [96]  0.159755513 -0.168294244 -0.056534304 -0.131507496 -0.419548996
#> 
#> Slot "Rp":
#>   [1] -0.94851072  1.49262807  0.44382044  0.55139158 -2.51033113 -0.05370066
#>   [7]  1.54537176  1.62785943  1.78974742  2.40455631 -2.25032182  5.52159001
#>  [13] -0.33701671 -0.51991735 -0.08807992 -1.53985599  1.88596630  0.07772469
#>  [19] -2.07987042  2.21565421 -1.30247413 -0.29795017  1.13028644  1.09184837
#>  [25]  0.51178216 -0.85188810  0.10760881 -5.94229182 -1.70483637 -1.36038607
#>  [31] -1.24301364  1.54336000 -0.93286606  0.89921070  0.93387218  1.35755340
#>  [37]  1.35557087 -1.00160062  0.60093930  2.02212024  3.59578570 -0.18213359
#>  [43]  2.71361574 -0.96812360 -0.05104588 -1.49102681 -0.92257660  1.26797009
#>  [49]  1.65642463 -1.28650029 -1.05596908 -1.67011827  3.52795361  0.37412793
#>  [55]  0.46626076 -1.98463027  0.41999643 -0.81243526  3.69278151 -1.04920060
#>  [61]  1.53519340 -0.32872344 -3.11674130  2.65491459 -5.26369203 -2.08383253
#>  [67] -1.12002844 -1.15166326  1.26743059 -0.35311636 -1.64496153 -2.58507592
#>  [73] -1.48325560 -1.31857599  1.61977275  0.52991418  3.62542179 -1.59425390
#>  [79]  0.15709546 -1.15881974 -0.98266114  0.11439466 -4.70454604  0.62498248
#>  [85]  1.03708365  2.91833611  1.77076770  2.48218046  0.26986969 -1.23951195
#>  [91]  3.26143634 -0.66779018 -0.26860787 -1.09524486 -1.41141509 -0.37783882
#>  [97] -0.95336421  0.43253740  1.62920938 -3.41352764
#> 
#> Slot "r2":
#> [1] 0.01149492
#> 

#data example taken from FAO Book
t0 <- c(3.2,0.8,0.2,2.6,4.4,5.2,2.7,1.2,4.5,5.6) # temperature 2 hours after sunset
tn <- c(-3.1,-5,-6.3,-5.4,-4,-2.5,-4.8,-5,-4.4,-3.3)
buildFAOTemp(temp=t0,tmin=tn)

#> An object of class "FAOFrostModel"
#> Slot "a":
#> [1] 0.5028121
#> 
#> Slot "b":
#> [1] 0
#> 
#> Slot "c":
#> [1] 0
#> 
#> Slot "Tp":
#>  [1] -4.299550 -5.506299 -5.807987 -4.601237 -3.696175 -3.293926 -4.550956
#>  [8] -5.305174 -3.645894 -3.092801
#> 
#> Slot "Rp":
#>  [1]  1.1995501  0.5062992 -0.4920135 -0.7987627 -0.3038245  0.7939258
#>  [7] -0.2490439  0.3051744 -0.7541057 -0.2071991
#> 
#> Slot "r2":
#> [1] 0.6656181
#>